requirements.txt
Django==4.2.4
django-allauth==0.55.2
django-apscheduler==0.6.2
django-filter==23.2
python-dotenv==1.0.0